Update API POST /{partnerId}/policies/{policyId}/credential-types-request #1861 (#1877)
* Update API POST /{partnerId}/policies/{policyId}/credential-types-request #1861


* test file fix


* fixed cr comments


* fixed cr comments


* test file fix


* test file fix


* test file fix


* file fix


* moved to partnermanagementservice


* fixed Test IMPl file


* fixed doc file


* added in helper class


* added in helper class


* fixed test file


* fixed test file


* fixed test file


* fixed test file


* fixed partnerservicecontroller bio-extrator changes


* fixed request body


* fixed impl file


---------
