+++ docs/partner_management_service_api_documentation.json
@@ -9969,13 +9969,13 @@
                 }
             }
         },
-        "/partners/{partnerId}/policies/{policyId}/credential-types-request": {
+        "/partner-policy-requests/{requestId}/credential-types-request": {
             "post": {
                 "tags": [
                     "partner-service-controller"
                 ],
                 "summary": "NEW: Submit credential type request",
-                "description": "Submits the credential type to associate with an **in-progress** partner policy mapping request. The value is stored for approval as part of the partner–policy mapping workflow.\r\n\r\n**Prerequisites**\r\n- A partner policy request must exist for `{partnerId}` with status **IN_PROGRESS**.\r\n- `request.partnerPolicyRequestId` must be the **business** request id (`pms.partner_policy_request.req_id`) for that mapping.\r\n- `{policyId}` in the path must match the policy id on that partner policy request.\r\n- Only **one** credential-type request is accepted per partner policy request.\r\n- The credential type must be one of the values configured for the deployment (`pmp.allowed.credential.types`, comma-separated; for example `auth`, `qrcode`, `euin`, `reprint` in a minimal setup).\r\n- The credential type must not already be mapped for this partner.\r\n\r\n**Success response:** HTTP 200 with `ResponseWrapper` envelope; `response` is the confirmation message string.\r\n\r\n**Deprecated replacement:** This endpoint replaces `POST /partners/{partnerId}/credentialtype/{credentialType}/policies/{policyName}` (deprecated since release-1.3.0-beta.5).\r\n\r\n**Authorization:** Uses `mosip.role.pms.postpartnersbioextractors` (same SpEL role key as submit bio extractors). Example role list in default configuration: PARTNER, CREDENTIAL_PARTNER, PARTNER_ADMIN, CREDENTIAL_ISSUANCE, CREATE_SHARE, ONLINE_VERIFICATION_PARTNER (verify against your deployment `application.properties`).",
+                "description": "Submits the credential type to associate with an **in-progress** partner policy mapping request. The value is stored for approval as part of the partner–policy mapping workflow.\r\n\r\n**Prerequisites**\r\n- A partner policy request must exist for `{requestId}` with status **IN_PROGRESS**.\r\n- `request.partnerPolicyRequestId` must be the **business** request id (`pms.partner_policy_request.req_id`) for that mapping.\r\n- `{policyId}` in the path must match the policy id on that partner policy request.\r\n- Only **one** credential-type request is accepted per partner policy request.\r\n- The credential type must be one of the values configured for the deployment (`pmp.allowed.credential.types`, comma-separated; for example `auth`, `qrcode`, `euin`, `reprint` in a minimal setup).\r\n- The credential type must not already be mapped for this partner.\r\n\r\n**Success response:** HTTP 200 with `ResponseWrapper` envelope; `response` is the confirmation message string.\r\n\r\n**Deprecated replacement:** This endpoint replaces `POST /partners/{partnerId}/credentialtype/{credentialType}/policies/{policyName}` (deprecated since release-1.3.0-beta.5).\r\n\r\n**Authorization:** Uses `mosip.role.pms.postpartnersbioextractors` (same SpEL role key as submit bio extractors). Example role list in default configuration: PARTNER, CREDENTIAL_PARTNER, PARTNER_ADMIN, CREDENTIAL_ISSUANCE, CREATE_SHARE, ONLINE_VERIFICATION_PARTNER (verify against your deployment `application.properties`).",
                 "operationId": "submitCredentialTypesRequest",
                 "parameters": [
                     {
@@ -9988,19 +9988,10 @@
                         }
                     },
                     {
-                        "name": "partnerId",
+                        "name": "requestId",
                         "in": "path",
                         "required": true,
-                        "description": "Partner id. Must match the partner that owns the partner policy request referenced in the body.",
-                        "schema": {
-                            "type": "string"
-                        }
-                    },
-                    {
-                        "name": "policyId",
-                        "in": "path",
-                        "required": true,
-                        "description": "Policy id. Must match the policy on the partner policy request referenced in the body.",
+                        "description": "Business partner-policy request id (`pms.partner_policy_request.req_id`).",
                         "schema": {
                             "type": "string"
                         }
@@ -10430,7 +10421,7 @@
                     "partner-management-controller"
                 ],
                 "summary": "NEW: Get credential type request for a partner-policy request",
-                "description": "Returns the credential type submitted for the partner policy mapping identified by `{requestId}`, if any.\r\n\r\n**Path parameter:** `{requestId}` is the **business** partner policy request identifier (`pms.partner_policy_request.req_id`), the same value used in `POST /partners/{partnerId}/policies/{policyId}/credential-types-request`.\r\n\r\n**Response shape:** `ResponseWrapperV2`-style envelope. Top-level `partnerPolicyRequestId` is the internal partner policy request key; `statusCode` reflects the partner policy request status. `response.credentialType` is the submitted credential type, or null if no credential-type request row exists yet.\r\n\r\n**Authorization:** `mosip.role.pms.getpartnersbioextractors` (same SpEL role key as the bio-extractors GET by partner-policy request). Example roles in default configuration: PARTNER, CREDENTIAL_PARTNER, PARTNER_ADMIN, CREDENTIAL_ISSUANCE, CREATE_SHARE, ONLINE_VERIFICATION_PARTNER.",
+                "description": "Returns the credential type submitted for the partner policy mapping identified by `{requestId}`, if any.\r\n\r\n**Path parameter:** `{requestId}` is the **business** partner policy request identifier (`pms.partner_policy_request.req_id`), the same value you send as `partnerPolicyRequestId` when calling `POST /partner-policy-requests/{requestId}/credential-types-request`.\r\n\r\n**Response shape:** `ResponseWrapperV2`-style envelope. Top-level `partnerPolicyRequestId` is the internal partner policy request key; `statusCode` reflects the partner policy request status. `response.credentialType` is the submitted credential type, or null if no credential-type request row exists yet.\r\n\r\n**Authorization:** `mosip.role.pms.getpartnersbioextractors` (same SpEL role key as the bio-extractors GET by partner-policy request). Example roles in default configuration: PARTNER, CREDENTIAL_PARTNER, PARTNER_ADMIN, CREDENTIAL_ISSUANCE, CREATE_SHARE, ONLINE_VERIFICATION_PARTNER.",
                 "operationId": "getPartnerPolicyRequestCredentialTypesRequest",
                 "parameters": [
                     {
