+++ partner/partner-management-service/src/main/java/io/mosip/pms/partner/manager/controller/PartnerManagementController.java
@@ -48,6 +48,7 @@ import io.mosip.pms.common.response.dto.ResponseWrapper;
 import io.mosip.pms.device.util.AuditUtil;
 import io.mosip.pms.partner.manager.constant.PartnerManageEnum;
 import io.mosip.pms.partner.manager.service.PartnerManagerService;
+import io.mosip.pms.partner.request.dto.CredentialTypeRequestDto;
 import io.mosip.pms.partner.request.dto.APIkeyStatusUpdateRequestDto;
 import io.mosip.pms.partner.response.dto.BioExtractorsResponseWrapperV2;
 import io.mosip.pms.partner.response.dto.CredentialTypesResponseWrapperV2;
@@ -76,7 +77,7 @@ public class PartnerManagementController {
 	
 	@Autowired
 	PartnerManagerService partnerManagementService;
-	
+
 	@Autowired
 	AuditUtil auditUtil;
 
@@ -109,6 +110,9 @@ public class PartnerManagementController {
 	
 	@Value("${mosip.pms.api.id.bioextractor.configuration.delete.patch}")
 	private String patchDeleteBioextractorConfigurationId;
+
+	@Value("${mosip.pms.api.id.partners.credentialtypes.request.post}")
+	private String postPartnerCredentialTypesRequestId;
 
 	String msg = "mosip.partnermanagement.partners.retrieve";
 	String version = "1.0";
@@ -546,6 +550,26 @@ public class PartnerManagementController {
 			@PathVariable("requestId") String requestId) {
 		inputValidator.validateRequestInput("requestId", requestId);
 		return partnerManagementService.getPartnerPolicyRequestCredentialTypes(requestId);
+	}
+
+	@PreAuthorize("hasAnyRole(@authorizedRoles.getPostpartnersbioextractors())")
+	@PostMapping(value = "/partner-policy-requests/{requestId}/credential-types-request")
+	@Operation(summary = "Service to submit credential types request",
+			description = "Persists credential type request against an in-progress partner policy mapping request. Partner and policy are resolved from the partner-policy request id in the path.")
+	public ResponseWrapperV2<String> submitCredentialTypesRequest(
+			@PathVariable("requestId") String requestId,
+			@RequestBody @Valid RequestWrapperV2<CredentialTypeRequestDto> requestWrapper) {
+		inputValidator.validateRequestInput("requestId", requestId);
+		Optional<ResponseWrapperV2<String>> validationResponse =
+				requestValidator.validate(postPartnerCredentialTypesRequestId, requestWrapper);
+		if (validationResponse.isPresent()) {
+			return validationResponse.get();
+		}
+		ResponseWrapperV2<String> response = new ResponseWrapperV2<>();
+		response.setResponse(partnerManagementService.submitCredentialTypesRequest(requestId,requestWrapper.getRequest()));
+		response.setId(requestWrapper.getId());
+		response.setVersion(requestWrapper.getVersion());
+		return response;
 	}
 
 	@Deprecated(since = "release-1.3.0-beta.2")
