+++ partner/partner-management-service/src/main/java/io/mosip/pms/partner/request/dto/CredentialTypeRequestDto.java
@@ -7,13 +7,6 @@ import lombok.Data;
 @Data
 public class CredentialTypeRequestDto {
 
-	/**
-	 * Partner policy request id ({@code pms.partner_policy_request.id}).
-	 */
-	@NotBlank(message = "partnerPolicyRequestId is missing")
-	@Size(max = 36, message = "partnerPolicyRequestId length must be <= 36")
-	private String partnerPolicyRequestId;
-
 	@NotBlank(message = "credentialType is missing")
 	@Size(max = 128, message = "credentialType length must be <= 128")
 	private String credentialType;
